only do pi thing on pi

+++ fixes.sh
@@ -2,6 +2,8 @@
 
 # No need to run this directly. Instead, run: sudo nems-update
 
+platform=$(/usr/local/share/nems/nems-scripts/info.sh platform)
+
 # NEMS 1.2.1 was released with an incorrect permission on this file
 chown www-data:www-data /etc/nagios3/global/timeperiods.cfg
 
@@ -202,10 +204,12 @@ if [ ! -f /var/log/nems/hw_model.log ]; then
 fi
 
 # Fix paths on rpimonitor
-if grep -q "/home/pi/nems-scripts/info.sh" /etc/rpimonitor/template/version.conf; then
-  systemctl stop rpimonitor
-  /bin/sed -i -- 's,/home/pi/nems-scripts/info.sh,/usr/local/bin/nems-info,g' /etc/rpimonitor/template/version.conf
-  systemctl start rpimonitor
+if (( $platform >= 0 )) && (( $platform <= 9 )); then
+  if grep -q "/home/pi/nems-scripts/info.sh" /etc/rpimonitor/template/version.conf; then
+    systemctl stop rpimonitor
+    /bin/sed -i -- 's,/home/pi/nems-scripts/info.sh,/usr/local/bin/nems-info,g' /etc/rpimonitor/template/version.conf
+    systemctl start rpimonitor
+  fi
 fi
 
 # Randomize nemsadmin password if NEMS is initialized
